Move i18n to shared folder
